♻️ :: modify basic exception package
