+++ domain/src/main/java/com/semicolon/domain/exception/Basic.kt
@@ -1,0 +1,19 @@
+package com.semicolon.domain.exception
+
+class BadRequestException : RuntimeException()
+
+class ConflictException : RuntimeException()
+
+class ForbiddenException : RuntimeException()
+
+class NoInternetException : RuntimeException()
+
+class NotFoundException : RuntimeException()
+
+class ServerException : RuntimeException()
+
+class TimeoutException : RuntimeException()
+
+class UnauthorizedException : RuntimeException()
+
+class UnknownException : RuntimeException()
