--- domain/src/main/java/com/semicolon/domain/exception/basic/NotFoundException.kt
@@ -1,3 +1,0 @@
-package com.semicolon.domain.exception.basic
-
-class NotFoundException : RuntimeException()
