+++ presentation/src/main/java/com/semicolon/walkhub/viewmodel/hub/HubMainViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.semicolon.domain.entity.rank.SchoolRankEntity
 import com.semicolon.domain.enums.DateType
-import com.semicolon.domain.exception.basic.NoInternetException
+import com.semicolon.domain.exception.NoInternetException
 import com.semicolon.domain.usecase.rank.FetchSchoolRankUseCase
 import com.semicolon.walkhub.ui.hub.model.HubSchoolRankData
 import com.semicolon.walkhub.util.MutableEventFlow
